examples/processor-verilog/pyproject.toml
[project]
name = "python-playground-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyserial",
]

[tool.autopep8]
indent_size = 2
